Fix for offset in const buffer not being set.

+++ Source/Runtime/UOEngine.Runtime.SDL3GPU/UO3DDxcCompiler.cs
@@ -143,7 +143,7 @@ internal class UOEngineDxcCompiler
                     shaderVariables[j] = new ShaderVariable
                     {
                         Name = varDesc.Name,
-                        Offset = varDesc.StartOffset,
+                        Offset = (int)varDesc.StartOffset,
                         Size = varDesc.Size,
                         Type = variableType
                     };
